+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - [10000](https://github.com/vegaprotocol/vega/issues/10000) - Introduce `cometbtf` command to replace `tendermint`.
 - [9948](https://github.com/vegaprotocol/vega/issues/9948) - Add support for linked stop orders.
 - [9849](https://github.com/vegaprotocol/vega/issues/9849) - Add database support for `num.Uint`.
+- [10275](https://github.com/vegaprotocol/vega/issues/10275) - Add API to list party's margin mode.
 
 ### 🐛 Fixes